src/requirements.txt
openai
gitpython
flask
python-dotenv
pinecone-io
beautifulsoup4
slimit
tqdm
libcst
